refactor(processor): inline v2 Process into main plugin files
Per maintainer review, the v1/v2 processor split only differs by
Process (v2) vs ProcessLogs (v1); a separate *_v2.go file per plugin
is unnecessary. Move each v2 Process method into its main plugin file,
mirroring flusher_sls (Flush + Export in one file), and drop the 19
*_v2.go implementation files. Test files keep the *_v2_test.go name.

No behavior change: the merged methods are unchanged passthrough
implementations.

Fixes #2603
