--- plugins/processor/filter/keyregex/processor_filter_key_regex_v2.go
@@ -1,29 +1,0 @@
-// Copyright 2026 iLogtail Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package keyregex
-
-import (
-	"github.com/alibaba/ilogtail/pkg/models"
-	"github.com/alibaba/ilogtail/pkg/pipeline"
-)
-
-// Process implements the v2 ProcessorV2 interface so this plugin can load in a
-// v2 (models.PipelineGroupEvents) pipeline. Filtering in v2 can drop events,
-// which is not implemented yet; to avoid silently dropping data it explicitly
-// passes all events (Log/Metric/Span) through unchanged rather than leaving v2
-// support undefined.
-func (p *ProcessorKeyFilter) Process(in *models.PipelineGroupEvents, context pipeline.PipelineContext) {
-	pipeline.CollectGroupEvents(context, in)
-}
