+++ plugins/processor/filter/regex/processor_filter_regex.go
@@ -18,6 +18,7 @@ import (
 	"regexp"
 
 	"github.com/alibaba/ilogtail/pkg/logger"
+	"github.com/alibaba/ilogtail/pkg/models"
 	"github.com/alibaba/ilogtail/pkg/pipeline"
 	"github.com/alibaba/ilogtail/pkg/protocol"
 	"github.com/alibaba/ilogtail/pkg/selfmonitor"
@@ -131,4 +132,13 @@ func init() {
 	pipeline.Processors[pluginType] = func() pipeline.Processor {
 		return &ProcessorRegexFilter{}
 	}
+}
+
+// Process implements the v2 ProcessorV2 interface so this plugin can load in a
+// v2 (models.PipelineGroupEvents) pipeline. Filtering in v2 can drop events,
+// which is not implemented yet; to avoid silently dropping data it explicitly
+// passes all events (Log/Metric/Span) through unchanged rather than leaving v2
+// support undefined.
+func (p *ProcessorRegexFilter) Process(in *models.PipelineGroupEvents, context pipeline.PipelineContext) {
+	pipeline.CollectGroupEvents(context, in)
 }
