+++ plugins/processor/grok/processor_grok.go
@@ -28,6 +28,7 @@ import (
 	"github.com/dlclark/regexp2"
 
 	"github.com/alibaba/ilogtail/pkg/logger"
+	"github.com/alibaba/ilogtail/pkg/models"
 	"github.com/alibaba/ilogtail/pkg/pipeline"
 	"github.com/alibaba/ilogtail/pkg/protocol"
 	"github.com/alibaba/ilogtail/pkg/selfmonitor"
@@ -446,4 +447,12 @@ func sortGraph(g graph) ([]string, []string) {
 		}
 	}
 	return order, nil
+}
+
+// Process implements the v2 ProcessorV2 interface so this plugin can load in a
+// v2 (models.PipelineGroupEvents) pipeline. It has no v2-native processing yet
+// and therefore explicitly passes all events (Log/Metric/Span) through
+// unchanged, rather than leaving v2 support undefined.
+func (p *ProcessorGrok) Process(in *models.PipelineGroupEvents, context pipeline.PipelineContext) {
+	pipeline.CollectGroupEvents(context, in)
 }
