+++ plugins/processor/strptime/strptime.go
@@ -23,6 +23,7 @@ import (
 	"github.com/knz/strtime"
 
 	"github.com/alibaba/ilogtail/pkg/logger"
+	"github.com/alibaba/ilogtail/pkg/models"
 	"github.com/alibaba/ilogtail/pkg/pipeline"
 	"github.com/alibaba/ilogtail/pkg/protocol"
 	"github.com/alibaba/ilogtail/pkg/selfmonitor"
@@ -192,4 +193,12 @@ func init() {
 	pipeline.Processors[pluginType] = func() pipeline.Processor {
 		return newStrptime()
 	}
+}
+
+// Process implements the v2 ProcessorV2 interface so this plugin can load in a
+// v2 (models.PipelineGroupEvents) pipeline. It has no v2-native processing yet
+// and therefore explicitly passes all events (Log/Metric/Span) through
+// unchanged, rather than leaving v2 support undefined.
+func (s *Strptime) Process(in *models.PipelineGroupEvents, context pipeline.PipelineContext) {
+	pipeline.CollectGroupEvents(context, in)
 }
